Fix Start-Timer implementation
Well it turns out that using a System.Threading.Thread's sleep function
inside a for loop was a really really bad idea because timeslice span it
takes to block a thread vary between different versions of Windows. In
addition to that it takes about 200,000 cycles to create a new thread
which in and of itself is a redundant operation if all do you is observing
differences in time. Therefore, the previous implementation added huge
delays after 10 seconds which is atrocious. Conversely, the StopWatch
object from the System.Diagnostic namespace is deemed suitable to accurately
measure elapsed time. After a few more test runs against my iPhone's timer
I have confirmed that this function now also works for count downs bigger
than ten seconds.